Move `addressParity` function to `Address.Gen` module.
This makes it reusable from other locations.

We also move `prop_addressParity_coverage` to `AddressSpec`.
